requirements.txt
asgiref==3.11.0
Django==6.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.3
pillow==12.1.0
python-decouple==3.8
sqlparse==0.5.5
